requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
html2text==2020.1.16
idna==2.10
isort==4.3.21
joblib==0.16.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.2
mccabe==0.6.1
numpy==1.19.0
pandas==1.0.5
pyenchant==3.1.1
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
scikit-learn==0.23.1
scipy==1.5.1
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
